+++ mobile/android/android-components/docs/changelog.md
@@ -4,13 +4,21 @@ title: Changelog
 permalink: /changelog/
 ---
 
-# 15.0.0-SNAPSHOT  (In Development)
+# 16.0.0-SNAPSHOT  (In Development)
 
-* [Commits](https://github.com/mozilla-mobile/android-components/compare/v14.0.0...master)
-* [Milestone](https://github.com/mozilla-mobile/android-components/milestone/75?closed=1)
+* [Commits](https://github.com/mozilla-mobile/android-components/compare/v15.0.0...master)
+* [Milestone](https://github.com/mozilla-mobile/android-components/milestone/76?closed=1)
 * [Dependencies](https://github.com/mozilla-mobile/android-components/blob/master/buildSrc/src/main/java/Dependencies.kt)
 * [Gecko](https://github.com/mozilla-mobile/android-components/blob/master/buildSrc/src/main/java/Gecko.kt)
 * [Configuration](https://github.com/mozilla-mobile/android-components/blob/master/buildSrc/src/main/java/Config.kt)
+
+# 15.0.0
+
+* [Commits](https://github.com/mozilla-mobile/android-components/compare/v14.0.0...v15.0.0)
+* [Milestone](https://github.com/mozilla-mobile/android-components/milestone/75?closed=1)
+* [Dependencies](https://github.com/mozilla-mobile/android-components/blob/v15.0.0/buildSrc/src/main/java/Dependencies.kt)
+* [Gecko](https://github.com/mozilla-mobile/android-components/blob/v15.0.0/buildSrc/src/main/java/Gecko.kt)
+* [Configuration](https://github.com/mozilla-mobile/android-components/blob/v15.0.0/buildSrc/src/main/java/Config.kt)
 
 * **browser-session**, **browser-state**, **feature-contextmenu**, **feature-downloads**
   * * ⚠️ **This is a breaking change**: Removed the `download` property from `Session`. Downloads can now only be observed on a `BrowserState` from the `browser-state` component. Therefore `ContextMenuUseCases` and `DownloadsUseCases` now require a `BrowserStore` instance.
@@ -22,6 +30,9 @@ permalink: /changelog/
 * **feature-customtabs**
   * Added `CustomTabWindowFeature` to handle windows inside custom tabs, PWAs, and TWAs.
 
+* **feature-tab-collections**
+
+  * Behavior change: In a collection List<TabEntity> is now ordered descending by creation date (newest tab in a collection on top)
 * **feature-session**, **engine-gecko-nightly** and **engine-gecko-beta**
   * Added api to manage the tracking protection exception list, any session added to the list will be ignored and the the current tracking policy will not be applied.
   ```kotlin